refactor: 컴포넌트 공통 요소 추출
Component.js, InlineComponent.js, PortalComponen.js 공통 로직 추출하여 BaseComponent에 저장
